switch to a mixin
