+++ index.js
@@ -1,4 +1,5 @@
 import './src/Toaster.js';
+import { ToasterMixin } from './src/ToasterMixin.js';
 import ToastEvent from './src/ToastEvent.js';
 
-export { ToastEvent };
+export { ToastEvent, ToasterMixin };
